+++ AINDY/routes/__init__.py
@@ -18,6 +18,7 @@ from .analytics_router import router as analytics_router
 from routes.genesis_router import router as genesis_router
 from routes.auth_router import router as auth_router
 from routes.masterplan_router import router as masterplan_router
+from routes.memory_router import router as memory_router
 
 
 ROUTERS = [
@@ -41,5 +42,6 @@ ROUTERS = [
     genesis_router,
     auth_router,
     masterplan_router,
+    memory_router,
 ]